Scrum 10 as a user i want to delete all of my information from friend s club so that i can be forgotten

+++ apps/post/src/application/commands/delete-post/delete-post.handler.ts
@@ -2,6 +2,7 @@ import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
 import { DeletePostCommand } from './delete-post.command';
 import { PostRepository } from '../../../infrastructure/post.db-entity.repository';
 import { PostDeletedEvent } from '../../events/post-deleted/post-deleted.event';
+import { UnauthorizedException } from '@nestjs/common';
 
 @CommandHandler(DeletePostCommand)
 export class DeletePostHandler implements ICommandHandler<DeletePostCommand> {
@@ -11,10 +12,15 @@ export class DeletePostHandler implements ICommandHandler<DeletePostCommand> {
   ) {}
 
   async execute({ deletePostRequest }: DeletePostCommand): Promise<any> {
-    const { postId } = deletePostRequest;
+    const { postId, userId } = deletePostRequest;
     const post = this.eventPublisher.mergeObjectContext(
       await this.postRepository.findOneById(postId),
     );
+
+    // to move to validator class:
+    if (post.getUserId() != userId) {
+      throw new UnauthorizedException('Unauthorized: User IDs do not match');
+    }
 
     await this.postRepository.remove(postId);
     post.apply(new PostDeletedEvent(postId));
